requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-tavily
langgraph
chromadb
pydantic
python-dotenv
fpdf2
markdown